Add option to specify an upload timeout
One issue with scripted dataset uploads from links has been that upload jobs
stuck in the running state (due to problems either on the Galaxy or the
remote side) would never be canceled by the code.
This would cause whole analysis batches to be stuck and the running jobs would
count towards the user's job concurrency limit (potentially even
freezing other analyses).

This changeset introduces a timeout for running upload jobs (default: 60
min) after which the upload job will be canceled and the upload be
reattempted (unless the maximum number of upload attempts is exceeded).