Make WF inputs findable in WF outputs history
If a history is populated as the result of a WF invocation, which uses
existing datasets in another history, these datasets will not always
become part of the new history. The change here will make such input
datasets robustly discoverable when searching the results history.

This fixes also the export bot in such cases because this bot needs to
discover three dataset collections from a consensus history, only one
of which is generated by the consensus WF, while the other two are WF
inputs.