Fix logic of summary generation, add new option
Introduces a new option --complete-only, which only retains records in
the summary json output, for which the Galaxy analysis has been completed
successfully. Makes sure that this, the pre-existing --retain-incomplete
option and the default behavior (retain only records, for which all
relevant histories got discovered, but do not check successful
completion) are applied to all output modes.