Add script to deduplicate tabular metadata
The script will only write the first three columns of lines on which any
sample first occurs.
When used on the sorted tabular output of the summary script, this means
that for any sample analyzed twice only the line representing the
analysis with the latest WF versions (and for ties, the one with the
latest completed date) will be kept.